requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
bcrypt
cryptography
cachetools
pydantic-settings
pytz